revoke licence
Former-commit-id: 079648cf36a48e1bbd571adeff3bbf2817732980 [formerly 0b3c730f4e782875c3ce24ad6226d182a725c3d8] [formerly aff07c94fea258d71fae4783f9a4b4c79122f669 [formerly 4be22ab9fa65c8d80922abe6f3ca9491ccd37859]]
Former-commit-id: 564c139e6aea9bdea287e613783c49de4b560172 [formerly a3e5c10b78c16ac0a7fb8e7816dc4c1c9fffcbe4]
Former-commit-id: 03cc1745676ffda2f0d3ca915fbd561cd46cd02d

+++ model.py
@@ -660,42 +660,43 @@ class Organization(ndb.Model):
         # init default stages,status, default values...
         cls.init_default_values(org_key)
 
+    def get_assigned_licenses(self):
+        return len(filter(lambda user: user.has_license(self.plan),
+                          User.fetch_by_organization(self.key)))
+
     @classmethod
     def assign_license(cls, org_key, user_key):
         organization = org_key.get()
         user = user_key.get()
-        if user.organization == org_key:
-            if user.license_status != 'active':
-                if organization.nb_used_licenses <= organization.nb_licenses:
-                    user.license_status = 'active'
-                    user.status = 'active'
-                    user.license_expires_on = organization.licenses_expires_on
-                    user.put()
-                    organization.nb_used_licenses += 1
-                    organization.put()
-                else:
-                    raise endpoints.UnauthorizedException('you need more licenses')
-        else:
-            raise endpoints.UnauthorizedException('the user is not withing your organization')
+        assigned_licenses = organization.get_assigned_licenses()
+        org_subscription = organization.get_subscription()
+        user_subscription = user.get_subscription()
+        if org_subscription.plan.get().name != config.PREMIUM:
+            raise endpoints.BadRequestException('You have to upgrade')
+        if org_subscription.quantity - assigned_licenses < 1:
+            raise endpoints.UnauthorizedException('you need more licenses')
+        if user_subscription.key == org_subscription.key:
+            raise endpoints.BadRequestException('this user already have an activated licences')
+        if user.organization != org_key:
+            raise endpoints.UnauthorizedException('The user is not withing your organization')
+        user.set_subscription(org_subscription)
+        # org_subscription.quantity -= 1
+        # org_subscription.put()
 
     @classmethod
     def unassign_license(cls, org_key, user_key):
         organization = org_key.get()
         user = user_key.get()
-        if user.organization == org_key:
-            print 'go ahead'
-            if user.license_status == 'active':
-                print 'active will be suspended'
-                user.status = 'suspended'
-                user.license_status = 'suspended'
-                user.license_expires_on = organization.licenses_expires_on
-                user.put()
-                organization.nb_used_licenses -= 1
-                organization.put()
-            else:
-                raise endpoints.UnauthorizedException('the user is already suspended')
-        else:
-            raise endpoints.UnauthorizedException('the user is not withing your organization')
+        org_subscription = organization.get_subscription()
+        user_subscription = user.get_subscription()
+        if user_subscription.plan.get().name != config.PREMIUM:
+            raise endpoints.BadRequestException('You have to upgrade')
+        if user_subscription.key != org_subscription.key:
+            raise endpoints.BadRequestException('this user already have an activated licences')
+        if user.organization != org_key:
+            raise endpoints.UnauthorizedException('The user is not withing your organization')
+        user.set_subscription(Subscription.create_freemium_subscription())
+
 
     @classmethod
     def upgrade_to_business_version(cls, org_key):
@@ -988,7 +989,8 @@ class User(EndpointsModel):
         if not isinstance(new_subscription, Subscription):
             raise ValueError('sub parameter should be of type {} '.format(Subscription.__class__))
         if self.subscription:
-            self.subscription.delete()
+            if self.subscription != self.organization.get().get_subscription().key:
+                self.subscription.delete()
         self.subscription = new_subscription.key
         self.put()
